Relocate all events to /events
Also pulls out event definitions from script loader into init.lua
Also deletes the empty controlscriptevents file
