+++ code/events/investigation.lua
