+++ code/scriptloader.lua
@@ -1,3 +1,8 @@
+require "code/events/core"
+require "code/events/init"
+require "code/events/trial"
+require "code/events/investigation"
+
 function LoadScript(scene, scriptPath)
     scene.events = {}
     scene.sceneScript = {}
@@ -323,137 +328,4 @@ function DisectLine(line)
     end
 
     return words
-end
-
-function NewAnimation(file, holdFirst)
-    local animation = {}
-    local source = love.graphics.newImage(file)
-
-    animation.source = source
-    animation.anim = {}
-
-    -- TODO: This is to support animation files that are configured for the DS res.
-    -- Once the files have been updated to have more white space between frames, should
-    -- be able to remove this.
-    local smallWidth = 256
-    if holdFirst then
-        for i=1, 20 do
-            animation.anim[i] = love.graphics.newQuad(0,0, smallWidth,GraphicsHeight(), source:getWidth(), source:getHeight())
-        end
-    end
-
-    for i=1, source:getWidth()/smallWidth do
-        local x = (i-1)*smallWidth
-        animation.anim[#animation.anim+1] = love.graphics.newQuad(x,0, smallWidth,GraphicsHeight(), source:getWidth(), source:getHeight())
-    end
-
-    return animation
-end
-
--- initializes all character files based on folder
-function NewCharInitEvent(name, location, gender)
-    local self = {}
-    self.name = name
-    self.gender = gender
-
-    -- allows for characters to be placed in characters/ or a custom directory
-    if string.match(location,"/") then
-        self.location = location
-    else
-        self.location = "characters/"..location
-    end
-
-    -- grabs the files in the character directory
-    self.files = love.filesystem.getDirectoryItems(self.location)
-
-    self.poses = {}
-    self.animations = {}
-    self.sounds = {}
-
-    -- sorts files by type and adds them to the scene
-    for b, i in ipairs(self.files) do
-        if string.match(i,".png") then
-
-            if string.match(i,"_ani") then
-                local a = i:gsub(".png","")
-                local a = a:gsub("_ani","")
-
-                self.animations[a] = NewAnimation(self.location.."/"..i, false)
-            elseif string.match(i,"_un") then
-                local a = i:gsub(".png","")
-                local a = a:gsub("_un","")
-
-                self.poses[a] = NewAnimation(self.location.."/"..i, false)
-            else
-                local a = i:gsub(".png","")
-                local isTalking = string.match(i, "Talking")
-
-                self.poses[a] = NewAnimation(self.location.."/"..i, not isTalking)
-            end
-
-        elseif string.match(i,".wav") then
-            local a = i:gsub(".wav","")
-            self.sounds[a] = love.audio.newSource(self.location.."/"..i, "static")
-            self.sounds[a]:setVolume(0.25)
-
-        elseif string.match(i,".WAV") then
-            local a = i:gsub(".WAV","")
-            self.sounds[a] = love.audio.newSource(self.location.."/"..i, "static")
-            self.sounds[a]:setVolume(0.25)
-        end
-    end
-
-    self.update = function (self, scene, dt)
-        scene.characters[self.name] = {
-            poses = self.poses,
-            animations = self.animations,
-            sounds = self.sounds,
-
-            location = self.location,
-            wideshot = NewAnimation(self.location .. "/wideshot.png", false),
-            name = self.name,
-            gender = self.gender,
-            frame = "Normal",
-        }
-
-        return false
-    end
-
-    return self
-end
-
-function NewEvidenceInitEvent(name, externalName, info, file)
-    local self = {}
-    self.name = name
-    self.externalName = externalName
-    self.info = info
-    self.file = file
-
-    self.update = function (self, scene, dt)
-        scene.evidence[self.name] = {
-            name = self.name,
-            externalName = self.externalName,
-            info = self.info,
-            sprite = love.graphics.newImage(self.file),
-        }
-
-        return false
-    end
-
-    return self
-end
-
-function AddToStack(events, sceneScript, event, lineParts)
-    -- This is used by the engine to see what event should
-    -- be run right now
-    table.insert(events, event)
-
-    -- Save just enough info for us to return to this
-    -- state if we choose to skip around the game
-    local eventContext = {
-        lineParts = lineParts,
-        event = event
-    }
-
-    table.insert(sceneScript, #events, eventContext)
 end
